requirements.txt
requests>=2.31.0
flask>=2.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
